requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contentful==2.4.0
contentful_management==2.14.4
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
